// demos/jsonpath-tools/src/modules/playground/models/custom-function.ts
import { Type } from "#components/jsonpath";

/**
 * Custom JSONPath function.
 */
export interface CustomFunction {
    /**
     * Name.
     */
    name: string;

    /**
     * Description. In Markdown format.
     */
    description: string;

    /**
     * List of function parameters.
     */
    parameters: readonly CustomFunctionParameter[];

    /**
     * Return type.
     */
    returnType: Type;

    /**
     * JavaScript code of the function implementation.
     */
    code: string;
}

/**
 * Parameter of a custom JSONPath function.
 */
export interface CustomFunctionParameter {
    /**
     * Name.
     */
    name: string;

    /**
     * Description. In Markdown format.
     */
    description: string;

    /**
     * Type.
     */
    type: Type;
}